fix: renamed `base` to `base_profiler_source`
